fix： 修复 note 和 category bug

+++ src/main/java/com/linmoblog/server/Entity/Category.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Category {
-    private long categoryKey;
+    private Integer categoryKey;
     private String categoryTitle;
     private String pathName;
     private String introduce;
